formatted all docs and removed unecessary console logs and comments

+++ controllers/userController.js
@@ -4,7 +4,7 @@ const userController = {
   async getUsers(req, res) {
     try {
       const users = await User.find();
-      res.status(200).json(users);
+      return res.status(200).json(users);
     } catch (err) {
       console.log(err);
       return res.status(500).json(err);
@@ -19,7 +19,7 @@ const userController = {
         return res.status(404).json({ message: "No user with that ID" });
       }
 
-      res.status(200).json(user);
+      return res.status(200).json(user);
     } catch (err) {
       console.log(err);
       return res.status(500).json(err);
@@ -29,7 +29,7 @@ const userController = {
   async createUser(req, res) {
     try {
       const user = await User.create(req.body);
-      res.status(200).json(user);
+      return res.status(200).json(user);
     } catch (err) {
       console.log(err);
       return res.status(500).json(err);
@@ -48,7 +48,7 @@ const userController = {
         return res.status(404).json({ message: "No user with this id!" });
       }
 
-      res.status(200).json(user);
+      return res.status(200).json(user);
     } catch (err) {
       console.log(err);
       return res.status(500).json(err);
@@ -57,36 +57,36 @@ const userController = {
 
   async deleteUser(req, res) {
     try {
-      const user = await User.findOneAndDelete({ _id: req.params.userId });
+      const user = await User.findOneAndDelete({ _id: req.params.userId }); //{ $pull: {thought: {_id: req.params.userId}} }
 
       if (!user) {
         return res.status(404).json({ message: "No user with that ID" });
       }
 
       await Thought.deleteMany({ _id: { $in: user.thoughts } });
-      res
+      return res
         .status(200)
-        .json({ message: "User and associated thoughts deleted!" });
+        .json({
+          message: "User and associated thoughts and reactions deleted!",
+        });
     } catch (err) {
       console.log(err);
       return res.status(500).json(err);
     }
   },
 
-  //   ???????????
   async addFriend(req, res) {
     try {
       const friend = await User.findOneAndUpdate(
         { _id: req.params.userId },
-        { $push: { friends: req.params.friendId } },
+        // add to set or pull???
+        { $addToSet: { friends: req.params.friendId } },
         { runValidators: true, new: true }
       );
 
-      // ?????
       if (!friend) {
         return res.status(404).json({ message: "No user with that ID" });
       }
-      // ?????
 
       return res.status(200).json(friend);
     } catch (err) {
@@ -94,24 +94,18 @@ const userController = {
       return res.status(500).json(err);
     }
   },
-  //   ????????????
 
-  //   ????????????????????????????????????????
   async deleteFriend(req, res) {
     try {
       const friend = await User.findOneAndUpdate(
         { _id: req.params.userId },
-        { $pull: { friends: { _id: req.params.friendId } } },
+        { $pull: { friends: req.params.friendId } },
         { runValidators: true, new: true }
       );
 
-      // ?????
       if (!friend) {
-        return res
-          .status(404)
-          .json({ message: "Check user and friend ID" });
+        return res.status(404).json({ message: "Check user and friend ID" });
       }
-      // ?????
 
       return res.status(200).json(friend);
     } catch (err) {
@@ -119,7 +113,6 @@ const userController = {
       return res.status(500).json(err);
     }
   },
-  // ??????????????????????????????????????????
 };
 
 module.exports = userController;
